+++ app/scripts/views/open-view.js
@@ -7,7 +7,6 @@ const SecureInput = require('../comp/secure-input');
 const DropboxChooser = require('../comp/dropbox-chooser');
 const KeyHandler = require('../comp/key-handler');
 const StorageFileListView = require('../views/storage-file-list-view');
-const FeatureDetector = require('../util/feature-detector');
 const Logger = require('../util/logger');
 const Locale = require('../util/locale');
 const UrlUtil = require('../util/url-util');
@@ -15,6 +14,7 @@ const InputFx = require('../util/input-fx');
 const Comparators = require('../util/comparators');
 const Storage = require('../storage');
 const Launcher = require('../comp/launcher');
+const FocusDetector = require('../comp/focus-detector');
 
 const logger = new Logger('open-view');
 
@@ -106,7 +106,9 @@ const OpenView = Backbone.View.extend({
     },
 
     focusInput: function() {
-        if (!FeatureDetector.isMobile) {
+        console.log('focus');
+        if (FocusDetector.hasFocus()) {
+            console.log('hasFocus');
             this.inputEl.focus();
         }
     },
@@ -378,20 +380,6 @@ const OpenView = Backbone.View.extend({
 
         const fileInfo = this.model.fileInfos.get(id);
         this.showOpenFileInfo(fileInfo);
-
-        if (fileInfo && Launcher && Launcher.fingerprints) {
-            this.openFileWithFingerprint(fileInfo);
-        }
-    },
-
-    openFileWithFingerprint(fileInfo) {
-        if (fileInfo.get('fingerprint')) {
-            Launcher.fingerprints.auth(fileInfo.id, fileInfo.get('fingerprint'), password => {
-                this.inputEl.val(password);
-                this.inputEl.trigger('input');
-                this.openDb();
-            });
-        }
     },
 
     removeFile: function(id) {
@@ -515,6 +503,8 @@ const OpenView = Backbone.View.extend({
         this.params.keyFileData = null;
         this.displayOpenFile();
         this.displayOpenKeyFile();
+
+        this.openFileWithFingerprint(fileInfo);
     },
 
     showOpenLocalFile: function(path, keyFilePath) {
@@ -534,6 +524,20 @@ const OpenView = Backbone.View.extend({
             this.params.keyFilePath = keyFilePath;
             this.params.keyFileData = null;
             this.displayOpenKeyFile();
+        }
+    },
+
+    openFileWithFingerprint: function(fileInfo) {
+        if (!fileInfo.has('fingerprint')) {
+            return;
+        }
+
+        if (Launcher && Launcher.fingerprints) {
+            Launcher.fingerprints.auth(fileInfo.id, fileInfo.get('fingerprint'), password => {
+                this.inputEl.val(password);
+                this.inputEl.trigger('input');
+                this.openDb();
+            });
         }
     },