+++ desktop/app.js
@@ -1,19 +1,23 @@
 const electron = require('electron');
-const app = electron.app;
-
-if (app.makeSingleInstance(restoreMainWindow)) app.quit();
-
 const path = require('path');
 const fs = require('fs');
 
+const app = electron.app;
+
 let mainWindow = null;
 let appIcon = null;
-let openFile = process.argv.filter(arg => /\.kdbx$/i.test(arg))[0];
 let ready = false;
 let appReady = false;
 let restartPending = false;
 let mainWindowPosition = {};
 let updateMainWindowPositionTimeout = null;
+
+const gotTheLock = app.requestSingleInstanceLock();
+if (!gotTheLock) {
+    app.quit();
+}
+
+let openFile = process.argv.filter(arg => /\.kdbx$/i.test(arg))[0];
 const userDataDir = app.getPath('userData').replace(/[\\/]temp[\\/]\d+\.\d+[\\/]?$/, '');
 const windowPositionFileName = path.join(userDataDir, 'window-position.json');
 const appSettingsFileName = path.join(userDataDir, 'app-settings.json');
@@ -30,6 +34,7 @@ const showDevToolsOnStart = process.argv.some(arg => arg.startsWith('--devtools'
 app.setPath('userData', path.join(tempUserDataPath, tempUserDataPathRand));
 
 setEnv();
+restorePreferences();
 
 app.on('window-all-closed', () => {
     if (restartPending) {
@@ -38,6 +43,7 @@ app.on('window-all-closed', () => {
         app.removeAllListeners('ready');
         app.removeAllListeners('open-file');
         app.removeAllListeners('activate');
+        app.removeAllListeners('second-instance');
         electron.globalShortcut.unregisterAll();
         electron.powerMonitor.removeAllListeners('suspend');
         electron.powerMonitor.removeAllListeners('resume');
@@ -78,6 +84,11 @@ app.on('activate', () => {
 });
 app.on('will-quit', () => {
     electron.globalShortcut.unregisterAll();
+});
+app.on('second-instance', () => {
+    if (mainWindow) {
+        restoreMainWindow();
+    }
 });
 app.restartApp = function () {
     restartPending = true;
@@ -152,7 +163,7 @@ function createMainWindow() {
     setMenu();
     mainWindow.loadURL(htmlPath);
     if (showDevToolsOnStart) {
-        mainWindow.openDevTools();
+        mainWindow.openDevTools({ mode: 'bottom' });
     }
     mainWindow.once('ready-to-show', () => {
         mainWindow.show();
@@ -162,6 +173,7 @@ function createMainWindow() {
     mainWindow.webContents.on('context-menu', onContextMenu);
     mainWindow.on('resize', delaySaveMainWindowPosition);
     mainWindow.on('move', delaySaveMainWindowPosition);
+    mainWindow.on('restore', coerceMainWindowPositionToConnectedDisplay);
     mainWindow.on('close', updateMainWindowPositionIfPending);
     mainWindow.on('blur', mainWindowBlur);
     mainWindow.on('closed', () => {
@@ -193,6 +205,7 @@ function restoreMainWindow() {
     }
     mainWindow.setSkipTaskbar(false);
     mainWindow.show();
+    coerceMainWindowPositionToConnectedDisplay();
     setTimeout(destroyAppIcon, 0);
 }
 
@@ -236,7 +249,6 @@ function updateMainWindowPosition() {
     }
     mainWindowPosition.maximized = mainWindow.isMaximized();
     mainWindowPosition.fullScreen = mainWindow.isFullScreen();
-    mainWindowPosition.displayBounds = require('electron').screen.getDisplayMatching(bounds).bounds;
     mainWindowPosition.changed = true;
 }
 
@@ -256,12 +268,8 @@ function restoreMainWindowPosition() {
             mainWindowPosition = JSON.parse(data);
             if (mainWindow && mainWindowPosition) {
                 if (mainWindowPosition.width && mainWindowPosition.height) {
-                    const displayBounds = require('electron').screen.getDisplayMatching(mainWindowPosition).bounds;
-                    const db = mainWindowPosition.displayBounds;
-                    if (displayBounds.x === db.x && displayBounds.y === db.y &&
-                        displayBounds.width === db.width && displayBounds.height === db.height) {
-                        mainWindow.setBounds(mainWindowPosition);
-                    }
+                    mainWindow.setBounds(mainWindowPosition);
+                    coerceMainWindowPositionToConnectedDisplay();
                 }
                 if (mainWindowPosition.maximized) { mainWindow.maximize(); }
                 if (mainWindowPosition.fullScreen) { mainWindow.setFullScreen(true); }
@@ -392,6 +400,28 @@ function setEnv() {
     }
 }
 
+function restorePreferences() {
+    const profileConfigPath = path.join(userDataDir, 'profile.json');
+
+    const newProfile = { dir: tempUserDataPathRand };
+    let oldProfile;
+    try {
+        oldProfile = JSON.parse(fs.readFileSync(profileConfigPath, 'utf8'));
+    } catch (e) { }
+
+    fs.writeFileSync(profileConfigPath, JSON.stringify(newProfile));
+
+    if (oldProfile && oldProfile.dir && /^[\d.]+$/.test(oldProfile.dir)) {
+        const oldProfilePath = path.join(tempUserDataPath, oldProfile.dir);
+        const newProfilePath = path.join(tempUserDataPath, newProfile.dir);
+        if (fs.existsSync(path.join(oldProfilePath, 'Cookies'))) {
+            fs.mkdirSync(newProfilePath);
+            fs.renameSync(path.join(oldProfilePath, 'Cookies'),
+                path.join(newProfilePath, 'Cookies'));
+        }
+    }
+}
+
 function deleteOldTempFiles() {
     if (app.oldTempFilesDeleted) {
         return;
@@ -426,7 +456,48 @@ function deleteRecursive(dir) {
 
 function hookRequestHeaders() {
     electron.session.defaultSession.webRequest.onBeforeSendHeaders((details, callback) => {
-        delete details.requestHeaders['Origin'];
+        if (!details.url.startsWith('ws:')) {
+            delete details.requestHeaders['Origin'];
+        }
         callback({cancel: false, requestHeaders: details.requestHeaders});
     });
+}
+
+// If a display is disconnected while KeeWeb is minimized, Electron does not
+// ensure that the restored window appears on a display that is still connected.
+// This checks to be sure the title bar is somewhere the user can grab it,
+// without making it impossible to minimize and restore a window keeping it
+// partially off-screen or straddling two displays if the user desires that.
+
+function coerceMainWindowPositionToConnectedDisplay() {
+    const eScreen = electron.screen;
+    const displays = eScreen.getAllDisplays();
+    if (!displays || !displays.length) return;
+    const windowBounds = mainWindow.getBounds();
+    const contentBounds = mainWindow.getContentBounds();
+    const tbLeft = windowBounds.x;
+    const tbRight = windowBounds.x + windowBounds.width;
+    const tbTop = windowBounds.y;
+    const tbBottom = contentBounds.y;
+    // 160px width and 2/3s the title bar height should be enough that the user can grab it
+    for (let i = 0; i < displays.length; ++i) {
+        const workArea = displays[i].workArea;
+        const overlapWidth = Math.min(tbRight, workArea.x + workArea.width) - Math.max(tbLeft, workArea.x);
+        const overlapHeight = Math.min(tbBottom, workArea.y + workArea.height) - Math.max(tbTop, workArea.y);
+        if (overlapWidth >= 160 && 3 * overlapHeight >= 2 * (tbBottom - tbTop)) return;
+    }
+    // If we get here, no display contains a big enough strip of the title bar
+    // that we can be confident the user can drag it into visibility.  Rather than
+    // attempt to guess what the user wants, just center it on the primary display.
+    // Try to keep the previous height and width, but clamp each to 90% of the workarea.
+    const workArea = eScreen.getPrimaryDisplay().workArea;
+    const newWidth = Math.min(windowBounds.width, Math.floor(0.9 * workArea.width));
+    const newHeight = Math.min(windowBounds.height, Math.floor(0.9 * workArea.height));
+    mainWindow.setBounds({
+        'x': workArea.x + Math.floor((workArea.width - newWidth) / 2),
+        'y': workArea.y + Math.floor((workArea.height - newHeight) / 2),
+        'width': newWidth,
+        'height': newHeight
+    });
+    updateMainWindowPosition();
 }